pyproject.toml
[project]
name = "fastapi-demo"
version = "1.0.0"
description = "零售業簡易ERP系統 - FastAPI Demo"
authors = [
    {name = "Demo Developer", email = "demo@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # FastAPI 核心
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",

    # 資料庫
    "sqlmodel>=0.0.22",
    "sqlalchemy>=2.0.36",
    "pymysql>=1.1.1",
    "alembic>=1.14.0",

    # Redis
    "redis>=5.2.0",

    # 認證
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.12",

    # Celery 背景任務
    "celery[redis]>=5.4.0",

    # 工具
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.28.0",

    # 日誌
    "loguru>=0.7.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "faker>=33.0.0",
    "black>=24.10.0",
    "isort>=5.13.0",
    "flake8>=7.1.0",
    "pre-commit>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ["py312"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
